requirements.txt
# Core Framework
Flask==3.0.0
Waitress==2.1.2
Flask-Session==0.5.0

# Database
pyodbc

# Authentication (Microsoft Entra ID)
msal==1.26.0
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Performance & Background Jobs (NEW)
Flask-Caching==2.1.0
Flask-APScheduler==1.13.1